refactor(errors): better organize top-level errors
Three main objectives in this commit:

- introduce a small module hierarchy for errors, so that it is possible
  to better navigate through the error types
- reduce code duplication by factoring out common patterns
- remove unnecessary error dynamism by utilizing existing error kind enums

We're finally beginning to approach a situation I would consider satisfactory
with regard to error handling:

- There is a top-level `error` module containing general-purpose crate-level errors
- It also contains specialty errors with which we can construct general-purpose wrapped error stacks
- Each code module group (i.e. folder) also contains its own internal error type, which
  contains unique leaf errors and sufficient information to wrap other crate error types.

I still need to analyze and eliminate leaf types which are never constructed,
fix the tests, and get rid of the legacy `CryptoError` entirely, but finally
the end of this work is in sight.
