+++ crypto/src/e2e_identity/enabled.rs
@@ -49,7 +49,7 @@ impl Client {
 
 #[cfg(test)]
 mod tests {
-    use crate::{e2e_identity::error::Error, prelude::MlsCredentialType, test_utils::*, CryptoError};
+    use crate::{e2e_identity::error::Error, prelude::MlsCredentialType, test_utils::*};
     use openmls_traits::types::SignatureScheme;
     use wasm_bindgen_test::*;
