+++ crypto/src/error/leaf.rs
@@ -1,0 +1,38 @@
+// Wire
+// Copyright (C) 2022 Wire Swiss GmbH
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+
+/// These errors can be raised from several different modules, so we centralize the definitions here
+/// to ease error-handling.
+#[derive(Debug, thiserror::Error)]
+pub enum LeafError {
+    /// This error is emitted when the requested conversation already exists with the given if
+    #[error("Conversation already exists")]
+    ConversationAlreadyExists(crate::prelude::ConversationId),
+    /// This error is emitted when the requested conversation couldn't be found in our store
+    #[error("Couldn't find conversation")]
+    ConversationNotFound(crate::prelude::ConversationId),
+    /// When looking for a X509 credential for a given ciphersuite and it has not been done
+    #[error("End-to-end identity enrollment has not been done")]
+    E2eiEnrollmentNotDone,
+    /// The MLS group is in an invalid state for an unknown reason
+    #[error("The MLS group is in an invalid state for an unknown reason")]
+    InternalMlsError,
+    /// Unexpectedly failed to retrieve group info
+    ///
+    /// This may be an implementation error.
+    #[error("unexpectedly failed to retrieve group info")]
+    MissingGroupInfo,
+}
